[refactor] Move "help" resources
Previously the methods to get a help embed were part of the "Info" module. This has been changed and the methods have been moved to CommandCollection and ModuleCollection and now accept the respective object's name as the first parameter. The privacy checks have also been moved to inside the help() methods. All calls to these methods were updated accordingly.