[structures] Remove need to duplicate Command's properties
Basically I don't have to type the same field out twice on both the Command class and the CommandOptions interface.
This does mean the I export it separately at the bottom of the file but that's kind of irrelevant (Idk, just thought I should mention it).
User 388834196132724756 on Discord helped me with this.