[refactor] Massive update
I tried to do individual commits for each thing that took place but it's such a royal pain in the ass that I just gave up. This doesn't matter, it's my project and nobody else will be working on this. Which means I am talking to myself right now. I should probably stop.