[refactor] Fix `dev` scripts running code twice at start
    - Added `tsc-watch` as dependency to `bot` and `web`
    - Removed `concurrently` module from `bot`'s dependencies (wasn't needed anymore)
Basically the typescript compiler (in watch mode) was making changes as `nodemon` was waiting for changes to reload the app, which caused the app to run, and then restart after a short pause. This is obviously not desirable. This is done via the `tsc-watch` module, which extends the typescript (watch) compiler by adding lifecycle commands; With which I can tell it to start the app *after* it successfully compiles